requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumbers==8.13.36
pillow==10.3.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
